Adds code coverage, fixes build
