+++ .github/workflows/build.yml
@@ -17,7 +17,7 @@ jobs:
       - name: Install Dependencies
         run: composer install --no-ansi --no-interaction --no-scripts --no-progress --prefer-dist
       - name: Lint
-        run: ./bin/phpcs
+        run: composer lint
   test:
     runs-on: ubuntu-latest
     strategy:
@@ -35,4 +35,11 @@ jobs:
       - name: Install Dependencies
         run: composer install --no-ansi --no-interaction --no-scripts --no-progress --prefer-dist
       - name: Test
-        run: ./bin/phpunit
+        run: API_EMAIL=mock_email@example.com API_KEY=123 composer coverage
+      - name: Coveralls
+        if: github.ref == 'refs/heads/main'
+        env:
+          COVERALLS_REPO_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        run: |
+          composer global require php-coveralls/php-coveralls
+          php-coveralls --coverage_clover=clover.xml -v
