+++ .gitignore
@@ -2,3 +2,5 @@ vendor
 bin
 .phplint-cache
 *.cache
+*.xml
+*.html
