+++ composer.json
@@ -28,14 +28,26 @@
         "squizlabs/php_codesniffer": "3.*",
         "php-vcr/php-vcr": "^1.5",
         "php-vcr/phpunit-testlistener-vcr": "^3.1",
-        "allejo/php-vcr-sanitizer": "^1.0"
+        "allejo/php-vcr-sanitizer": "^1.0.9",
+        "php-coveralls/php-coveralls": "^2.4"
     },
     "config": {
         "bin-dir": "bin"
     },
+    "scripts": {
+        "coverage": "XDEBUG_MODE=coverage ./bin/phpunit --coverage-html clover.html --coverage-clover clover.xml",
+        "fix": "./bin/phpcbf",
+        "lint": "./bin/phpcs",
+        "test": "./bin/phpunit"
+    },
     "autoload": {
-        "psr-0": {
-            "TuneupTechnology": "src/"
+        "psr-4": {
+            "TuneupTechnology\\": "src/"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "TuneupTechnology\\Tests\\": "test/TuneupTechnology/"
         }
     },
     "version": "2.0.0"
