+++ src/TuneupTechnology.php
@@ -4,7 +4,7 @@
 
 // Import the client library files
 require_once(dirname(__FILE__) . '/TuneupTechnology/Client.php');
-require_once(dirname(__FILE__) . '/TuneupTechnology/Customer.php');
+require_once(dirname(__FILE__) . '/TuneupTechnology/Customers.php');
 require_once(dirname(__FILE__) . '/TuneupTechnology/Inventory.php');
-require_once(dirname(__FILE__) . '/TuneupTechnology/Location.php');
-require_once(dirname(__FILE__) . '/TuneupTechnology/Ticket.php');
+require_once(dirname(__FILE__) . '/TuneupTechnology/Locations.php');
+require_once(dirname(__FILE__) . '/TuneupTechnology/Tickets.php');
