+++ test/cassettes/customers/retrieve.yml
@@ -18,13 +18,13 @@
         headers:
             Cache-Control: 'no-cache, private'
             Content-Type: application/json
-            Date: 'Wed, 19 May 2021 05:32:55 GMT'
+            Date: 'Sun, 23 May 2021 05:05:52 GMT'
             Server: nginx
             X-Powered-By: PHP/7.4.15
             X-Ratelimit-Limit: '60'
             X-Ratelimit-Remaining: '58'
-            Content-Length: '345'
-        body: '{"data":{"id":1,"firstname":"Mariane McDermott","lastname":"Della Legros","email":"hyatt.rene@example.org","phone":"8018981234","user_id":1,"notes":"Rerum eius rerum architecto molestias earum.","created_at":"2021-05-19 05:31:45","updated_at":"2021-05-19 05:31:45","deleted_at":null,"location_id":1},"message":"Customer retrieved successfully."}'
+            Content-Length: '363'
+        body: '{"data":{"id":1,"firstname":"Miss Clarabelle Lang","lastname":"Earl Wunsch III","email":"flavie.langworth@example.com","phone":"8018981234","user_id":1,"notes":"Et facilis ratione maxime aliquam nisi voluptatem.","created_at":"2021-05-23 05:05:17","updated_at":"2021-05-23 05:05:17","deleted_at":null,"location_id":1},"message":"Customer retrieved successfully."}'
         curl_info:
             url: 'http://tuneapp.localhost/api/customers/1'
             content_type: application/json
@@ -34,32 +34,32 @@
             filetime: -1
             ssl_verify_result: 0
             redirect_count: 0
-            total_time: 0.131146
-            namelookup_time: 0.000637
-            connect_time: 0.000747
-            pretransfer_time: 0.000772
+            total_time: 0.115347
+            namelookup_time: 0.000594
+            connect_time: 0.000707
+            pretransfer_time: 0.000729
             size_upload: !!float 0
-            size_download: !!float 345
-            speed_download: !!float 2633
+            size_download: !!float 363
+            speed_download: !!float 3156
             speed_upload: !!float 0
-            download_content_length: !!float 345
+            download_content_length: !!float 363
             upload_content_length: !!float 0
-            starttransfer_time: 0.131122
+            starttransfer_time: 0.11533
             redirect_time: !!float 0
             redirect_url: ''
             primary_ip: 127.0.0.1
             certinfo: {  }
             primary_port: 80
             local_ip: 127.0.0.1
-            local_port: 51731
+            local_port: 49951
             http_version: 2
             protocol: 1
             ssl_verifyresult: 0
             scheme: HTTP
             appconnect_time_us: 0
-            connect_time_us: 747
-            namelookup_time_us: 637
-            pretransfer_time_us: 772
+            connect_time_us: 707
+            namelookup_time_us: 594
+            pretransfer_time_us: 729
             redirect_time_us: 0
-            starttransfer_time_us: 131122
-            total_time_us: 131146
+            starttransfer_time_us: 115330
+            total_time_us: 115347
