+++ test/cassettes/customers/update.yml
@@ -21,13 +21,13 @@
         headers:
             Cache-Control: 'no-cache, private'
             Content-Type: application/json
-            Date: 'Wed, 19 May 2021 05:32:56 GMT'
+            Date: 'Sun, 23 May 2021 05:05:52 GMT'
             Server: nginx
             X-Powered-By: PHP/7.4.15
             X-Ratelimit-Limit: '60'
             X-Ratelimit-Remaining: '56'
             Content-Length: '307'
-        body: '{"data":{"id":1,"firstname":"Jake","lastname":"Peralta","email":"jake@example.com","phone":"8015551234","user_id":1,"notes":"Believes he is a good detective.","created_at":"2021-05-19 05:31:45","updated_at":"2021-05-19 05:32:56","deleted_at":null,"location_id":2},"message":"Customer updated successfully."}'
+        body: '{"data":{"id":1,"firstname":"Jake","lastname":"Peralta","email":"jake@example.com","phone":"8015551234","user_id":1,"notes":"Believes he is a good detective.","created_at":"2021-05-23 05:05:17","updated_at":"2021-05-23 05:05:52","deleted_at":null,"location_id":2},"message":"Customer updated successfully."}'
         curl_info:
             url: 'http://tuneapp.localhost/api/customers/1'
             content_type: application/json
@@ -37,32 +37,32 @@
             filetime: -1
             ssl_verify_result: 0
             redirect_count: 0
-            total_time: 0.125652
-            namelookup_time: 0.000595
-            connect_time: 0.000704
-            pretransfer_time: 0.000728
+            total_time: 0.115007
+            namelookup_time: 0.000626
+            connect_time: 0.000776
+            pretransfer_time: 0.000814
             size_upload: !!float 160
             size_download: !!float 307
-            speed_download: !!float 2456
-            speed_upload: !!float 1280
+            speed_download: !!float 2669
+            speed_upload: !!float 1391
             download_content_length: !!float 307
             upload_content_length: !!float 160
-            starttransfer_time: 0.125636
+            starttransfer_time: 0.114988
             redirect_time: !!float 0
             redirect_url: ''
             primary_ip: 127.0.0.1
             certinfo: {  }
             primary_port: 80
             local_ip: 127.0.0.1
-            local_port: 51733
+            local_port: 49953
             http_version: 2
             protocol: 1
             ssl_verifyresult: 0
             scheme: HTTP
             appconnect_time_us: 0
-            connect_time_us: 704
-            namelookup_time_us: 595
-            pretransfer_time_us: 728
+            connect_time_us: 776
+            namelookup_time_us: 626
+            pretransfer_time_us: 814
             redirect_time_us: 0
-            starttransfer_time_us: 125636
-            total_time_us: 125652
+            starttransfer_time_us: 114988
+            total_time_us: 115007
