+++ test/cassettes/inventory/delete.yml
@@ -18,13 +18,13 @@
         headers:
             Cache-Control: 'no-cache, private'
             Content-Type: application/json
-            Date: 'Wed, 19 May 2021 05:32:56 GMT'
+            Date: 'Sun, 23 May 2021 05:05:53 GMT'
             Server: nginx
             X-Powered-By: PHP/7.4.15
             X-Ratelimit-Limit: '60'
             X-Ratelimit-Remaining: '50'
             Content-Length: '328'
-        body: '{"data":{"id":1,"name":"Inventory Item","quantity":"1","created_at":"2021-05-19 05:31:45","updated_at":"2021-05-19 05:32:56","deleted_at":"2021-05-19 05:32:56","inventory_type_id":1,"part_number":"1234","sku":"1234","notes":"here are some notes","part_price":"19.99","location_id":1},"message":"Inventory deleted successfully."}'
+        body: '{"data":{"id":1,"name":"Inventory Item","quantity":"1","created_at":"2021-05-23 05:05:17","updated_at":"2021-05-23 05:05:53","deleted_at":"2021-05-23 05:05:53","inventory_type_id":1,"part_number":"1234","sku":"1234","notes":"here are some notes","part_price":"19.99","location_id":1},"message":"Inventory deleted successfully."}'
         curl_info:
             url: 'http://tuneapp.localhost/api/inventory/1'
             content_type: application/json
@@ -34,32 +34,32 @@
             filetime: -1
             ssl_verify_result: 0
             redirect_count: 0
-            total_time: 0.135947
-            namelookup_time: 0.00061
-            connect_time: 0.000715
-            pretransfer_time: 0.000735
+            total_time: 0.1139
+            namelookup_time: 0.000641
+            connect_time: 0.000748
+            pretransfer_time: 0.000768
             size_upload: !!float 0
             size_download: !!float 328
-            speed_download: !!float 2429
+            speed_download: !!float 2902
             speed_upload: !!float 0
             download_content_length: !!float 328
             upload_content_length: !!float 0
-            starttransfer_time: 0.135924
+            starttransfer_time: 0.113879
             redirect_time: !!float 0
             redirect_url: ''
             primary_ip: 127.0.0.1
             certinfo: {  }
             primary_port: 80
             local_ip: 127.0.0.1
-            local_port: 51739
+            local_port: 49959
             http_version: 2
             protocol: 1
             ssl_verifyresult: 0
             scheme: HTTP
             appconnect_time_us: 0
-            connect_time_us: 715
-            namelookup_time_us: 610
-            pretransfer_time_us: 735
+            connect_time_us: 748
+            namelookup_time_us: 641
+            pretransfer_time_us: 768
             redirect_time_us: 0
-            starttransfer_time_us: 135924
-            total_time_us: 135947
+            starttransfer_time_us: 113879
+            total_time_us: 113900
