+++ test/cassettes/inventory/retrieve.yml
@@ -18,13 +18,13 @@
         headers:
             Cache-Control: 'no-cache, private'
             Content-Type: application/json
-            Date: 'Wed, 19 May 2021 05:32:56 GMT'
+            Date: 'Sun, 23 May 2021 05:05:53 GMT'
             Server: nginx
             X-Powered-By: PHP/7.4.15
             X-Ratelimit-Limit: '60'
             X-Ratelimit-Remaining: '53'
-            Content-Length: '359'
-        body: '{"data":{"id":1,"name":"enim","quantity":"5","created_at":"2021-05-19 05:31:45","updated_at":"2021-05-19 05:31:45","deleted_at":null,"inventory_type_id":1,"part_number":"531232","sku":"785766","notes":"Dignissimos temporibus provident earum reiciendis et modi voluptates.","part_price":null,"location_id":1},"message":"Inventory item retrieved successfully."}'
+            Content-Length: '341'
+        body: '{"data":{"id":1,"name":"et","quantity":"7","created_at":"2021-05-23 05:05:17","updated_at":"2021-05-23 05:05:17","deleted_at":null,"inventory_type_id":1,"part_number":"237091","sku":"133367","notes":"Quos ipsam nihil est eveniet tempora fuga voluptatem.","part_price":null,"location_id":1},"message":"Inventory item retrieved successfully."}'
         curl_info:
             url: 'http://tuneapp.localhost/api/inventory/1'
             content_type: application/json
@@ -34,32 +34,32 @@
             filetime: -1
             ssl_verify_result: 0
             redirect_count: 0
-            total_time: 0.136474
-            namelookup_time: 0.000751
-            connect_time: 0.000883
-            pretransfer_time: 0.000912
+            total_time: 0.109709
+            namelookup_time: 0.000637
+            connect_time: 0.000771
+            pretransfer_time: 0.000798
             size_upload: !!float 0
-            size_download: !!float 359
-            speed_download: !!float 2639
+            size_download: !!float 341
+            speed_download: !!float 3128
             speed_upload: !!float 0
-            download_content_length: !!float 359
+            download_content_length: !!float 341
             upload_content_length: !!float 0
-            starttransfer_time: 0.136459
+            starttransfer_time: 0.109693
             redirect_time: !!float 0
             redirect_url: ''
             primary_ip: 127.0.0.1
             certinfo: {  }
             primary_port: 80
             local_ip: 127.0.0.1
-            local_port: 51736
+            local_port: 49956
             http_version: 2
             protocol: 1
             ssl_verifyresult: 0
             scheme: HTTP
             appconnect_time_us: 0
-            connect_time_us: 883
-            namelookup_time_us: 751
-            pretransfer_time_us: 912
+            connect_time_us: 771
+            namelookup_time_us: 637
+            pretransfer_time_us: 798
             redirect_time_us: 0
-            starttransfer_time_us: 136459
-            total_time_us: 136474
+            starttransfer_time_us: 109693
+            total_time_us: 109709
