+++ test/cassettes/inventory/update.yml
@@ -21,13 +21,13 @@
         headers:
             Cache-Control: 'no-cache, private'
             Content-Type: application/json
-            Date: 'Wed, 19 May 2021 05:32:56 GMT'
+            Date: 'Sun, 23 May 2021 05:05:53 GMT'
             Server: nginx
             X-Powered-By: PHP/7.4.15
             X-Ratelimit-Limit: '60'
             X-Ratelimit-Remaining: '51'
             Content-Length: '307'
-        body: '{"data":{"id":1,"name":"Inventory Item","quantity":1,"created_at":"2021-05-19 05:31:45","updated_at":"2021-05-19 05:32:56","deleted_at":null,"inventory_type_id":1,"part_number":"1234","sku":"1234","notes":"here are some notes","part_price":19.99,"location_id":1},"message":"Inventory updated successfully."}'
+        body: '{"data":{"id":1,"name":"Inventory Item","quantity":1,"created_at":"2021-05-23 05:05:17","updated_at":"2021-05-23 05:05:53","deleted_at":null,"inventory_type_id":1,"part_number":"1234","sku":"1234","notes":"here are some notes","part_price":19.99,"location_id":1},"message":"Inventory updated successfully."}'
         curl_info:
             url: 'http://tuneapp.localhost/api/inventory/1'
             content_type: application/json
@@ -37,32 +37,32 @@
             filetime: -1
             ssl_verify_result: 0
             redirect_count: 0
-            total_time: 0.120274
-            namelookup_time: 0.000682
-            connect_time: 0.000784
-            pretransfer_time: 0.000807
+            total_time: 0.113254
+            namelookup_time: 0.000624
+            connect_time: 0.000733
+            pretransfer_time: 0.000755
             size_upload: !!float 159
             size_download: !!float 307
-            speed_download: !!float 2558
-            speed_upload: !!float 1325
+            speed_download: !!float 2716
+            speed_upload: !!float 1407
             download_content_length: !!float 307
             upload_content_length: !!float 159
-            starttransfer_time: 0.120259
+            starttransfer_time: 0.11324
             redirect_time: !!float 0
             redirect_url: ''
             primary_ip: 127.0.0.1
             certinfo: {  }
             primary_port: 80
             local_ip: 127.0.0.1
-            local_port: 51738
+            local_port: 49958
             http_version: 2
             protocol: 1
             ssl_verifyresult: 0
             scheme: HTTP
             appconnect_time_us: 0
-            connect_time_us: 784
-            namelookup_time_us: 682
-            pretransfer_time_us: 807
+            connect_time_us: 733
+            namelookup_time_us: 624
+            pretransfer_time_us: 755
             redirect_time_us: 0
-            starttransfer_time_us: 120259
-            total_time_us: 120274
+            starttransfer_time_us: 113240
+            total_time_us: 113254
