+++ test/cassettes/locations/retrieve.yml
@@ -18,13 +18,13 @@
         headers:
             Cache-Control: 'no-cache, private'
             Content-Type: application/json
-            Date: 'Wed, 19 May 2021 05:32:57 GMT'
+            Date: 'Sun, 23 May 2021 05:05:53 GMT'
             Server: nginx
             X-Powered-By: PHP/7.4.15
             X-Ratelimit-Limit: '60'
             X-Ratelimit-Remaining: '48'
-            Content-Length: '267'
-        body: '{"data":{"id":1,"name":"Kozey, O''Reilly and Kuphal","street":"86888 Cassin Dam","city":"Adalbertoborough","state":"IA","zip":"67524","created_at":"2021-05-19 05:31:45","updated_at":"2021-05-19 05:31:45","deleted_at":null},"message":"Location retrieved successfully."}'
+            Content-Length: '264'
+        body: '{"data":{"id":1,"name":"McCullough-Ryan","street":"11154 Joanne Road Apt. 187","city":"New Kavonburgh","state":"TN","zip":"26086","created_at":"2021-05-23 05:05:17","updated_at":"2021-05-23 05:05:17","deleted_at":null},"message":"Location retrieved successfully."}'
         curl_info:
             url: 'http://tuneapp.localhost/api/locations/1'
             content_type: application/json
@@ -34,32 +34,32 @@
             filetime: -1
             ssl_verify_result: 0
             redirect_count: 0
-            total_time: 0.120841
-            namelookup_time: 0.000604
-            connect_time: 0.000722
-            pretransfer_time: 0.000746
+            total_time: 0.105688
+            namelookup_time: 0.00073
+            connect_time: 0.000855
+            pretransfer_time: 0.00088
             size_upload: !!float 0
-            size_download: !!float 267
-            speed_download: !!float 2225
+            size_download: !!float 264
+            speed_download: !!float 2514
             speed_upload: !!float 0
-            download_content_length: !!float 267
+            download_content_length: !!float 264
             upload_content_length: !!float 0
-            starttransfer_time: 0.120822
+            starttransfer_time: 0.105666
             redirect_time: !!float 0
             redirect_url: ''
             primary_ip: 127.0.0.1
             certinfo: {  }
             primary_port: 80
             local_ip: 127.0.0.1
-            local_port: 51741
+            local_port: 49962
             http_version: 2
             protocol: 1
             ssl_verifyresult: 0
             scheme: HTTP
             appconnect_time_us: 0
-            connect_time_us: 722
-            namelookup_time_us: 604
-            pretransfer_time_us: 746
+            connect_time_us: 855
+            namelookup_time_us: 730
+            pretransfer_time_us: 880
             redirect_time_us: 0
-            starttransfer_time_us: 120822
-            total_time_us: 120841
+            starttransfer_time_us: 105666
+            total_time_us: 105688
