+++ test/cassettes/locations/update.yml
@@ -21,13 +21,13 @@
         headers:
             Cache-Control: 'no-cache, private'
             Content-Type: application/json
-            Date: 'Wed, 19 May 2021 05:32:57 GMT'
+            Date: 'Sun, 23 May 2021 05:05:54 GMT'
             Server: nginx
             X-Powered-By: PHP/7.4.15
             X-Ratelimit-Limit: '60'
             X-Ratelimit-Remaining: '46'
             Content-Length: '245'
-        body: '{"data":{"id":1,"name":"Location Name","street":"123 California Ave","city":"Salt Lake","state":"UT","zip":84043,"created_at":"2021-05-19 05:31:45","updated_at":"2021-05-19 05:32:57","deleted_at":null},"message":"Location updated successfully."}'
+        body: '{"data":{"id":1,"name":"Location Name","street":"123 California Ave","city":"Salt Lake","state":"UT","zip":84043,"created_at":"2021-05-23 05:05:17","updated_at":"2021-05-23 05:05:54","deleted_at":null},"message":"Location updated successfully."}'
         curl_info:
             url: 'http://tuneapp.localhost/api/locations/1'
             content_type: application/json
@@ -37,32 +37,32 @@
             filetime: -1
             ssl_verify_result: 0
             redirect_count: 0
-            total_time: 0.121214
-            namelookup_time: 0.000644
-            connect_time: 0.000755
-            pretransfer_time: 0.000779
+            total_time: 0.108522
+            namelookup_time: 0.000622
+            connect_time: 0.000731
+            pretransfer_time: 0.000755
             size_upload: !!float 98
             size_download: !!float 245
-            speed_download: !!float 2024
-            speed_upload: !!float 809
+            speed_download: !!float 2268
+            speed_upload: !!float 907
             download_content_length: !!float 245
             upload_content_length: !!float 98
-            starttransfer_time: 0.121198
+            starttransfer_time: 0.108507
             redirect_time: !!float 0
             redirect_url: ''
             primary_ip: 127.0.0.1
             certinfo: {  }
             primary_port: 80
             local_ip: 127.0.0.1
-            local_port: 51743
+            local_port: 49964
             http_version: 2
             protocol: 1
             ssl_verifyresult: 0
             scheme: HTTP
             appconnect_time_us: 0
-            connect_time_us: 755
-            namelookup_time_us: 644
-            pretransfer_time_us: 779
+            connect_time_us: 731
+            namelookup_time_us: 622
+            pretransfer_time_us: 755
             redirect_time_us: 0
-            starttransfer_time_us: 121198
-            total_time_us: 121214
+            starttransfer_time_us: 108507
+            total_time_us: 108522
