+++ test/cassettes/tickets/all.yml
@@ -18,13 +18,13 @@
         headers:
             Cache-Control: 'no-cache, private'
             Content-Type: application/json
-            Date: 'Wed, 19 May 2021 05:32:57 GMT'
+            Date: 'Sun, 23 May 2021 05:05:55 GMT'
             Server: nginx
             X-Powered-By: PHP/7.4.15
             X-Ratelimit-Limit: '60'
             X-Ratelimit-Remaining: '42'
             Transfer-Encoding: chunked
-        body: '{"data":{"current_page":1,"data":[{"id":1,"customer_id":1,"ticket_type_id":1,"serial":"169909","user_id":1,"notes":"Iste animi voluptas perferendis odio iure consectetur animi ab.","title":"Asperiores non ea adipisci est voluptas.","status":1,"created_at":"2021-05-19 05:31:45","updated_at":"2021-05-19 05:31:45","deleted_at":null,"device":"iPhone 12","imei":"234967","location_id":1},{"id":2,"customer_id":1,"ticket_type_id":1,"serial":"330356","user_id":1,"notes":"Sequi cum magni aut consequuntur vitae nulla.","title":"Libero consectetur quaerat suscipit.","status":1,"created_at":"2021-05-19 05:31:45","updated_at":"2021-05-19 05:31:45","deleted_at":null,"device":"iPhone 12","imei":"446852","location_id":1},{"id":3,"customer_id":1,"ticket_type_id":1,"serial":"461776","user_id":1,"notes":"Voluptatem sit nihil illo distinctio expedita similique.","title":"Fugit maxime doloremque delectus voluptatem vero.","status":1,"created_at":"2021-05-19 05:31:45","updated_at":"2021-05-19 05:31:45","deleted_at":null,"device":"iPhone 12","imei":"333388","location_id":1},{"id":4,"customer_id":1,"ticket_type_id":1,"serial":"151019","user_id":1,"notes":"Veniam molestiae consectetur sed nesciunt inventore.","title":"Quo in non minima.","status":1,"created_at":"2021-05-19 05:31:45","updated_at":"2021-05-19 05:31:45","deleted_at":null,"device":"iPhone 12","imei":"233233","location_id":1},{"id":5,"customer_id":1,"ticket_type_id":1,"serial":"757006","user_id":1,"notes":"Corrupti accusamus hic eum explicabo quia quis officiis.","title":"Autem enim eveniet odio.","status":1,"created_at":"2021-05-19 05:31:45","updated_at":"2021-05-19 05:31:45","deleted_at":null,"device":"iPhone 12","imei":"359101","location_id":1},{"id":6,"customer_id":2,"ticket_type_id":1,"serial":"10000","user_id":1,"notes":"here are some notes","title":"Fancy Title","status":1,"created_at":"2021-05-19 05:32:57","updated_at":"2021-05-19 05:32:57","deleted_at":null,"device":"2","imei":"10000","location_id":2}],"first_page_url":"http:\/\/tuneapp.localhost\/api\/tickets?page=1","from":1,"last_page":1,"last_page_url":"http:\/\/tuneapp.localhost\/api\/tickets?page=1","next_page_url":null,"path":"http:\/\/tuneapp.localhost\/api\/tickets","per_page":20,"prev_page_url":null,"to":6,"total":6},"message":"Tickets retrieved successfully."}'
+        body: '{"data":{"current_page":1,"data":[{"id":1,"customer_id":1,"ticket_type_id":1,"serial":"839270","user_id":1,"notes":"Dolor quos non doloribus.","title":"Ipsa et nihil atque expedita quidem.","status":1,"created_at":"2021-05-23 05:05:17","updated_at":"2021-05-23 05:05:17","deleted_at":null,"device":"iPhone 12","imei":"945819","location_id":1},{"id":2,"customer_id":1,"ticket_type_id":1,"serial":"592920","user_id":1,"notes":"Delectus odio fugiat facilis et dolores nostrum corporis.","title":"Aliquam voluptate rerum vero qui quo et ducimus enim.","status":1,"created_at":"2021-05-23 05:05:17","updated_at":"2021-05-23 05:05:17","deleted_at":null,"device":"iPhone 12","imei":"142573","location_id":1},{"id":3,"customer_id":1,"ticket_type_id":1,"serial":"803486","user_id":1,"notes":"Et quo qui perspiciatis reprehenderit dolorem earum.","title":"Explicabo quod rerum nemo dicta occaecati ad incidunt.","status":1,"created_at":"2021-05-23 05:05:17","updated_at":"2021-05-23 05:05:17","deleted_at":null,"device":"iPhone 12","imei":"951207","location_id":1},{"id":4,"customer_id":1,"ticket_type_id":1,"serial":"768967","user_id":1,"notes":"Aut itaque qui asperiores tenetur voluptas praesentium.","title":"Veritatis dolorem delectus architecto dolore.","status":1,"created_at":"2021-05-23 05:05:17","updated_at":"2021-05-23 05:05:17","deleted_at":null,"device":"iPhone 12","imei":"188006","location_id":1},{"id":5,"customer_id":1,"ticket_type_id":1,"serial":"350010","user_id":1,"notes":"Nihil optio distinctio a cum repellendus autem consequuntur.","title":"Necessitatibus qui mollitia sint aliquam nihil.","status":1,"created_at":"2021-05-23 05:05:17","updated_at":"2021-05-23 05:05:17","deleted_at":null,"device":"iPhone 12","imei":"543072","location_id":1},{"id":6,"customer_id":2,"ticket_type_id":1,"serial":"10000","user_id":1,"notes":"here are some notes","title":"Fancy Title","status":1,"created_at":"2021-05-23 05:05:54","updated_at":"2021-05-23 05:05:54","deleted_at":null,"device":"2","imei":"10000","location_id":2}],"first_page_url":"http:\/\/tuneapp.localhost\/api\/tickets?page=1","from":1,"last_page":1,"last_page_url":"http:\/\/tuneapp.localhost\/api\/tickets?page=1","next_page_url":null,"path":"http:\/\/tuneapp.localhost\/api\/tickets","per_page":20,"prev_page_url":null,"to":6,"total":6},"message":"Tickets retrieved successfully."}'
         curl_info:
             url: 'http://tuneapp.localhost/api/tickets'
             content_type: application/json
@@ -34,32 +34,32 @@
             filetime: -1
             ssl_verify_result: 0
             redirect_count: 0
-            total_time: 0.117865
-            namelookup_time: 0.000573
-            connect_time: 0.000683
-            pretransfer_time: 0.000708
+            total_time: 0.106599
+            namelookup_time: 0.000607
+            connect_time: 0.000714
+            pretransfer_time: 0.000734
             size_upload: !!float 0
-            size_download: !!float 2306
-            speed_download: !!float 19709
+            size_download: !!float 2351
+            speed_download: !!float 22179
             speed_upload: !!float 0
             download_content_length: !!float -1
             upload_content_length: !!float 0
-            starttransfer_time: 0.117848
+            starttransfer_time: 0.106581
             redirect_time: !!float 0
             redirect_url: ''
             primary_ip: 127.0.0.1
             certinfo: {  }
             primary_port: 80
             local_ip: 127.0.0.1
-            local_port: 51747
+            local_port: 49968
             http_version: 2
             protocol: 1
             ssl_verifyresult: 0
             scheme: HTTP
             appconnect_time_us: 0
-            connect_time_us: 683
-            namelookup_time_us: 573
-            pretransfer_time_us: 708
+            connect_time_us: 714
+            namelookup_time_us: 607
+            pretransfer_time_us: 734
             redirect_time_us: 0
-            starttransfer_time_us: 117848
-            total_time_us: 117865
+            starttransfer_time_us: 106581
+            total_time_us: 106599
