+++ test/cassettes/tickets/delete.yml
@@ -18,13 +18,13 @@
         headers:
             Cache-Control: 'no-cache, private'
             Content-Type: application/json
-            Date: 'Wed, 19 May 2021 05:32:58 GMT'
+            Date: 'Sun, 23 May 2021 05:05:55 GMT'
             Server: nginx
             X-Powered-By: PHP/7.4.15
             X-Ratelimit-Limit: '60'
             X-Ratelimit-Remaining: '40'
             Content-Length: '334'
-        body: '{"data":{"id":1,"customer_id":2,"ticket_type_id":1,"serial":"10000","user_id":1,"notes":"here are some notes","title":"Fancy Title","status":1,"created_at":"2021-05-19 05:31:45","updated_at":"2021-05-19 05:32:58","deleted_at":"2021-05-19 05:32:58","device":"2","imei":"10000","location_id":2},"message":"Ticket deleted successfully."}'
+        body: '{"data":{"id":1,"customer_id":2,"ticket_type_id":1,"serial":"10000","user_id":1,"notes":"here are some notes","title":"Fancy Title","status":1,"created_at":"2021-05-23 05:05:17","updated_at":"2021-05-23 05:05:55","deleted_at":"2021-05-23 05:05:55","device":"2","imei":"10000","location_id":2},"message":"Ticket deleted successfully."}'
         curl_info:
             url: 'http://tuneapp.localhost/api/tickets/1'
             content_type: application/json
@@ -34,32 +34,32 @@
             filetime: -1
             ssl_verify_result: 0
             redirect_count: 0
-            total_time: 0.57288
-            namelookup_time: 0.000655
-            connect_time: 0.000759
-            pretransfer_time: 0.000779
+            total_time: 0.105975
+            namelookup_time: 0.000561
+            connect_time: 0.000673
+            pretransfer_time: 0.000694
             size_upload: !!float 0
             size_download: !!float 334
-            speed_download: !!float 583
+            speed_download: !!float 3180
             speed_upload: !!float 0
             download_content_length: !!float 334
             upload_content_length: !!float 0
-            starttransfer_time: 0.572865
+            starttransfer_time: 0.105956
             redirect_time: !!float 0
             redirect_url: ''
             primary_ip: 127.0.0.1
             certinfo: {  }
             primary_port: 80
             local_ip: 127.0.0.1
-            local_port: 51749
+            local_port: 49970
             http_version: 2
             protocol: 1
             ssl_verifyresult: 0
             scheme: HTTP
             appconnect_time_us: 0
-            connect_time_us: 759
-            namelookup_time_us: 655
-            pretransfer_time_us: 779
+            connect_time_us: 673
+            namelookup_time_us: 561
+            pretransfer_time_us: 694
             redirect_time_us: 0
-            starttransfer_time_us: 572865
-            total_time_us: 572880
+            starttransfer_time_us: 105956
+            total_time_us: 105975
