+++ test/cassettes/tickets/update.yml
@@ -21,13 +21,13 @@
         headers:
             Cache-Control: 'no-cache, private'
             Content-Type: application/json
-            Date: 'Wed, 19 May 2021 05:32:58 GMT'
+            Date: 'Sun, 23 May 2021 05:05:55 GMT'
             Server: nginx
             X-Powered-By: PHP/7.4.15
             X-Ratelimit-Limit: '60'
             X-Ratelimit-Remaining: '41'
             Content-Length: '315'
-        body: '{"data":{"id":1,"customer_id":2,"ticket_type_id":1,"serial":"10000","user_id":1,"notes":"here are some notes","title":"Fancy Title","status":1,"created_at":"2021-05-19 05:31:45","updated_at":"2021-05-19 05:32:58","deleted_at":null,"device":"2","imei":10000,"location_id":2},"message":"Ticket updated successfully."}'
+        body: '{"data":{"id":1,"customer_id":2,"ticket_type_id":1,"serial":"10000","user_id":1,"notes":"here are some notes","title":"Fancy Title","status":1,"created_at":"2021-05-23 05:05:17","updated_at":"2021-05-23 05:05:55","deleted_at":null,"device":"2","imei":10000,"location_id":2},"message":"Ticket updated successfully."}'
         curl_info:
             url: 'http://tuneapp.localhost/api/tickets/1'
             content_type: application/json
@@ -37,32 +37,32 @@
             filetime: -1
             ssl_verify_result: 0
             redirect_count: 0
-            total_time: 0.11623
-            namelookup_time: 0.000606
-            connect_time: 0.000716
-            pretransfer_time: 0.00074
+            total_time: 0.114206
+            namelookup_time: 0.000648
+            connect_time: 0.000768
+            pretransfer_time: 0.000794
             size_upload: !!float 170
             size_download: !!float 315
-            speed_download: !!float 2715
-            speed_upload: !!float 1465
+            speed_download: !!float 2763
+            speed_upload: !!float 1491
             download_content_length: !!float 315
             upload_content_length: !!float 170
-            starttransfer_time: 0.116215
+            starttransfer_time: 0.114189
             redirect_time: !!float 0
             redirect_url: ''
             primary_ip: 127.0.0.1
             certinfo: {  }
             primary_port: 80
             local_ip: 127.0.0.1
-            local_port: 51748
+            local_port: 49969
             http_version: 2
             protocol: 1
             ssl_verifyresult: 0
             scheme: HTTP
             appconnect_time_us: 0
-            connect_time_us: 716
-            namelookup_time_us: 606
-            pretransfer_time_us: 740
+            connect_time_us: 768
+            namelookup_time_us: 648
+            pretransfer_time_us: 794
             redirect_time_us: 0
-            starttransfer_time_us: 116215
-            total_time_us: 116230
+            starttransfer_time_us: 114189
+            total_time_us: 114206
